requirements.txt
asgiref==3.7.2
certifi==2023.7.22
Django==4.2.5
django-debug-toolbar==4.2.0
psycopg==3.1.12
psycopg-binary==3.1.12
sentry-sdk==1.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.5
